refactor: Remove unused components and styles, reorganize file structure
