+++ mission-fe/src/App.jsx
@@ -1,5 +1,5 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Home from "./pages/Home";
+import Home from "./pages/home/Home";
 import Products from "./pages/Products";
 import Registration from "./components/features/registration/Registration";
