+++ mission-fe/src/pages/Products.jsx
@@ -1,5 +1,5 @@
-import Header from "../components/layout/Header";
-import Footer from "../components/layout/Footer";
+import Header from "../components/layout/header/Header";
+import Footer from "../components/layout/footer/Footer";
 import BestItems from "../components/features/products/BestItems";
 import OnSaleItems from "../components/features/products/OnSaleItems";
